resturant_system/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0

# Requests for API Gateway communication
requests==2.31.0

# For JWT Authentication (Auth service)
djangorestframework-simplejwt==5.3.1

# RabbitMQ communication 
pika==1.3.2

# Consul client (service discovery)
python-consul==1.1.0

# Deployment Utils
gunicorn==21.2.0
